# eb-backend/ip-manager-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ISP IP Management & Infrastructure API",
    "version": "1.0.0",
    "description": "Production-grade enterprise core IPAM and hardware inventory engine. Fully relational, audited, and protected via authorization matrices."
  },
  "servers": [
    {
      "url": "http://10.249.2.9"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid token to access infrastructure components. Obtain this via /api/ip-manager/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/ip-manager/vpns": {
      "get": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Get secure network virtual private listings",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Create new record in IP Tunnels & VPNs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/ip-manager/users": {
      "get": {
        "summary": "List all IP Manager users",
        "responses": {
          "200": {
            "description": "List of users with permissions"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a real login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "editor",
                      "viewer"
                    ]
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/ip-manager/users/{id}": {
      "patch": {
        "summary": "Update a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "perms": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/ip-manager/vpns/{id}": {
      "patch": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Update a record in IP Tunnels & VPNs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Delete a record in IP Tunnels & VPNs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}